Move commenting setting from Details to Security settings
