+++ app/scenes/Settings/Security.tsx
@@ -5,7 +5,7 @@ import { useState } from "react";
 import * as React from "react";
 import { useTranslation, Trans } from "react-i18next";
 import { toast } from "sonner";
-import { TeamPreference, EmailDisplay } from "@shared/types";
+import { CommentingAccess, TeamPreference, EmailDisplay } from "@shared/types";
 import ConfirmationDialog from "~/components/ConfirmationDialog";
 import Heading from "~/components/Heading";
 import type { Option } from "~/components/InputSelect";
@@ -66,6 +66,28 @@ function Security() {
           type: "item",
           label: t("No one"),
           value: EmailDisplay.None,
+        },
+      ] satisfies Option[],
+    [t]
+  );
+
+  const commentingOptions: Option[] = React.useMemo(
+    () =>
+      [
+        {
+          type: "item",
+          label: t("Members"),
+          value: CommentingAccess.Members,
+        },
+        {
+          type: "item",
+          label: t("Members and guests"),
+          value: CommentingAccess.Everyone,
+        },
+        {
+          type: "item",
+          label: t("No one"),
+          value: CommentingAccess.None,
         },
       ] satisfies Option[],
     [t]
@@ -165,6 +187,17 @@ function Security() {
       const preferences = {
         ...team.preferences,
         [TeamPreference.EmailDisplay]: emailDisplay,
+      };
+      await saveData({ preferences });
+    },
+    [saveData, team.preferences]
+  );
+
+  const handleCommentingChange = React.useCallback(
+    async (commenting: string) => {
+      const preferences = {
+        ...team.preferences,
+        [TeamPreference.Commenting]: commenting,
       };
       await saveData({ preferences });
     },
@@ -331,6 +364,23 @@ function Security() {
           options={emailDisplayOptions}
           onChange={handleEmailDisplayChange}
           label={t("Email address visibility")}
+          labelHidden
+          short
+        />
+      </SettingRow>
+      <SettingRow
+        label={t("Commenting")}
+        name={TeamPreference.Commenting}
+        description={t("Controls who can add comments to documents")}
+      >
+        <InputSelect
+          value={
+            team.getPreference(TeamPreference.Commenting) ||
+            CommentingAccess.Members
+          }
+          options={commentingOptions}
+          onChange={handleCommentingChange}
+          label={t("Commenting")}
           labelHidden
           short
         />
