+++ shared/i18n/locales/en_US/translation.json
@@ -1249,8 +1249,6 @@
   "Custom emojis can be used throughout your workspace in documents, comments, and reactions.": "Custom emojis can be used throughout your workspace in documents, comments, and reactions.",
   "Left": "Left",
   "Right": "Right",
-  "Members and guests": "Members and guests",
-  "No one": "No one",
   "Logo updated": "Logo updated",
   "Unable to upload new logo": "Unable to upload new logo",
   "Delete workspace": "Delete workspace",
@@ -1275,7 +1273,6 @@
   "This is the screen that workspace members will first see when they sign in.": "This is the screen that workspace members will first see when they sign in.",
   "Separate editing": "Separate editing",
   "When enabled documents have a separate editing mode by default instead of being always editable. This setting can be overridden by user preferences.": "When enabled documents have a separate editing mode by default instead of being always editable. This setting can be overridden by user preferences.",
-  "Controls who can add comments to documents.": "Controls who can add comments to documents.",
   "Danger": "Danger",
   "You can delete this entire workspace including collections, documents, and users.": "You can delete this entire workspace including collections, documents, and users.",
   "Enabled": "Enabled",
@@ -1371,6 +1368,8 @@
   "Photo": "Photo",
   "Choose a photo or image to represent yourself.": "Choose a photo or image to represent yourself.",
   "This could be your real name, or a nickname — however you’d like people to refer to you.": "This could be your real name, or a nickname — however you’d like people to refer to you.",
+  "Members and guests": "Members and guests",
+  "No one": "No one",
   "Are you sure you want to require invites?": "Are you sure you want to require invites?",
   "New users will first need to be invited to create an account. <em>Default role</em> and <em>Allowed domains</em> will no longer apply.": "New users will first need to be invited to create an account. <em>Default role</em> and <em>Allowed domains</em> will no longer apply.",
   "Settings that impact the access, security, and content of your workspace.": "Settings that impact the access, security, and content of your workspace.",
@@ -1389,6 +1388,7 @@
   "When enabled, users can delete their own account from the workspace": "When enabled, users can delete their own account from the workspace",
   "Email address visibility": "Email address visibility",
   "Controls who can see user email addresses in the workspace": "Controls who can see user email addresses in the workspace",
+  "Controls who can add comments to documents": "Controls who can add comments to documents",
   "Collection creation": "Collection creation",
   "Allow editors to create new collections within the workspace": "Allow editors to create new collections within the workspace",
   "Workspace creation": "Workspace creation",
